requirements.txt
backcall==0.2.0
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
future==0.18.2
imageio==2.9.0
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
networkx==2.6.2
numexpr==2.7.3
numpy==1.21.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pyFFTW==0.12.0
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.2
PyWavelets==1.1.1
pyzmq==22.1.0
scikit-image==0.18.2
scipy==1.7.0
six==1.16.0
sporco==0.1.12
tifffile==2021.7.30
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
wcwidth==0.2.5
